tensorflow_impl/requirements.txt
tensorflow==2.3.1
numpy==1.19.1
grpcio==1.33.2
grpcio-tools==1.33.2
tensorflow_datasets==4.0.1